Fix/windows tests (#44)
* Remove nnls patch and fix scipy version (with fixed nnls)

* ✨ remove patched nnls (using fixed scipy version now)

* 💄 linting; line breaks

* 🚑 fix dtypes to ensure windows ci to pass

---------

+++ tests/unit_tests/test_tabular_data.py
@@ -186,9 +186,9 @@ def test_dataframe_reader(psm_df_6):
         reader.read(), pd.DataFrame({"test": [1, 2, 3]})
     )
 
-    reader = DataFrameReader.from_array(np.array([1, 2, 3]), name="test")
+    reader = DataFrameReader.from_array(np.array([1, 2, 3], dtype=np.int32), name="test")
     pd.testing.assert_frame_equal(
-        reader.read(), pd.DataFrame({"test": [1, 2, 3]})
+        reader.read(), pd.DataFrame({"test": [1, 2, 3]}, dtype=np.int32)
     )
